// packages/core/src/@types/challengermode.d.ts
export interface GQLMatchResults {
  /**
   * Whether the results are final and will propagate within the competition.
   */
  final: boolean;

  /**
   * Whether the match ended in a draw.
   */
  draw: boolean;

  /**
   * The results of the each of lineups in the match.
   */
  lineupResults: Array<GQLMatchResult>;
}
// ...
export interface GQLStream {
  /**
   * The broadcast that this stream is a part of.
   */
  broadcast: GQLBroadcast;
  type: GQLStreamType;
}
// ...
export interface GQLCompetitionContext {
  /**
   * If the context is within a Tournament, this property will provide the tournament context.
   */
  tournamentContext?: GQLITournamentContext;
}
// ...
export interface GQLGameSpecificSettings {
  /**
   * JSON stored as an escaped string value holding game-title specific game settings.
   * @deprecated Use JsonValue instead.
   */
  value: string;
}
// ...
export interface GQLHosts {
  /**
   * List of Challengermode space hosts.
   */
  spaces: Array<GQLSpace>;
}
// ...
export interface GQLMatch {
  /**
   * The list of lineups in the match.
   */
  lineups: Array<GQLMatchLineup>;

  /**
   * The results of the match.
   */
  results: GQLMatchResults;

  /**
   * The tournament context of the match, if the match is part of a tournament.
   * @deprecated Use CompetitionContext instead.
   */
  tournamentContext?: GQLTournamentMatchContext;

  /**
   * The unique identifier of the match.
   */
  id: GQLUUID;

  /**
   * The current state of the match.
   */
  state: GQLMatchState;

  /**
   * The public name of the match.
   */
  name: string;

  /**
   * Numbers of minutes the players have to go to the Challengermode match lobby after the match starts.
   * If set to 0 the match will start without players going to the match screen.
   */
  maximumLobbyMinutes: number;

  /**
   * If not set to 0, players must join the in-game lobby during this period to avoid losing on walkover.
   */
  maximumGoToGameMinutes: number;

  /**
   * Information about how to find a a game session, including it's in-game game session ID.
   * @deprecated Use GameSession instead.
   */
  gameSessionTag: GQLGameSessionTag;

  /**
   * The in-game game session corresponding to this match.
   */
  gameSession: GQLGameSession;

  /**
   * Statistics for this match, including game session statistics.
   */
  statistics: GQLMatchStatisticsCollection;

  /**
   * The competition context of the match.
   */
  context: GQLCompetitionContext;
}

/**
 * A lineup is one or many players playing together in this match.
 */
export interface GQLMatchLineup {
  /**
   * The members in this match lineup.
   */
  members: Array<GQLMatchMember>;

  /**
   * Get statistics corresponding to this lineup.
   */
  statistics?: GQLStatisticsConnection;

  /**
   * The lineup number, 0-indexed.
   */
  number: number;

  /**
   * Whether the lineup missed the ready check.
   */
  missedReadyCheck: boolean;

  /**
   * The lineup name.
   */
  lineupName: string;

  /**
   * If determined, the resulting position of the lineup in the match.
   * @deprecated Use MatchResult instead.
   */
  position?: number;

  /**
   * If determined, the resulting score of the lineup in the match.
   * @deprecated Use MatchResult instead.
   */
  score?: number;
}
// ...
export interface GQLMatchSeries {
  /**
   * Ordered list of matches in this match series.
   */
  matches: Array<GQLMatch>;

  /**
   * Specific game settings used for this match series. This can differ from the default game settings configured on the tournament.
   * @deprecated Use GameSessionSettings instead.
   */
  gameSettings?: GQLGameSpecificSettings;

  /**
   * Specific game settings used for this match series. This can differ from the default game settings configured on the tournament.
   * @deprecated Use game specific settings instead.
   */
  gameSessionSettings?: GQLJSON;

  /**
   * The tournament that this match series is a part of.
   * @deprecated Use Context instead.
   */
  tournament: GQLTournament;

  /**
   * The competition context of the match series.
   */
  context: GQLCompetitionContext;

  /**
   * List of lineups in the match series.
   */
  lineups: Array<GQLTournamentLineup | null>;

  /**
   * The results of the match series.
   */
  results: GQLMatchResults;
  title: string;
  labels: Array<GQLTournamentNodeLabel>;

  /**
   * The streams covering this series.
   */
  streams: Array<GQLStream>;
  id: GQLUUID;

  /**
   * The ordinal number of the match series in the tournament.
   */
  ordinal: number;

  /**
   * The current state of the match series.
   */
  state: GQLMatchSeriesState;

  /**
   * The number of lineups in each match of the match series.
   */
  lineupCount: number;

  /**
   * The time when the match series started. Null if the match series has not started yet.
   */
  startedAt?: GQLDateTime;

  /**
   * The best-of setting used for this match series. For example, "best of 3" means that the first team to win 2 matches wins the match series.
   */
  bestOf: number;

  /**
   * The scheduled start time of the first match of the series. May be null.
   */
  scheduledStartTimeAt?: GQLDateTime;
}
// ...
export interface GQLTournament {
  /**
   * A description of the current tournament.
   */
  description: string;

  /**
   * The stages of this tournament, played consecutively. This provides the overall structure of the tournament.
   */
  stages: Array<GQLTournamentStage>;

  /**
   * All match series that are a part of the tournament.
   * @deprecated Use `stages` instead.
   */
  matchSeries: Array<GQLMatchSeries>;

  /**
   * List of the Spaces hosting this tournament.
   */
  hosts: GQLHosts;

  /**
   * Provides an overview of participation in a tournament, including details of the players and teams involved, the roster, and the waiting list.
   */
  attendance: GQLTournamentAttendance;

  /**
   * Key timestamps related to the tournament, such as the start time and the opening of the confirmation window.
   */
  schedule: GQLTournamentSchedule;

  /**
   * Lists restrictions that users must meet to be eligible to participate in the competition.
   */
  competitionRestrictions: GQLCompetitionRestrictions;

  /**
   * List of restrictions for joining the tournament.
   * @deprecated Use competitionRestrictions instead.
   */
  restrictions: GQLTournamentRestrictions;

  /**
   * Links to external resources related to the tournament.
   */
  links: GQLTournamentLinks;

  /**
   * The participation status of the current user.
   */
  ownParticipation?: GQLOwnTournamentParticipation;

  /**
   * The streams covering this tournament.
   */
  streams: Array<GQLStream>;

  /**
   * Get which game title this tournament is for.
   */
  gameTitle: GQLGameTitle;

  /**
   * The tournament's unique ID.
   */
  id: GQLUUID;

  /**
   * The name of the tournament.
   */
  name?: string;

  /**
   * Current state of the tournament.
   */
  state: GQLTournamentState;

  /**
   * The contact address for the organiser.
   */
  contactUrl: string;

  /**
   * Settings for how the current tournament and matches are run.
   */
  settings: GQLTournamentSettingsGroup;
  idSuffix?: string;
}
// ...
export interface GQLTournamentLineup {
  /**
   * The team logo for this lineup. This will be available primarily if the lineup has a registered Team on challengermode.
   */
  logo?: GQLImage;

  /**
   * The team banner image.
   */
  banner?: GQLImage;

  /**
   * Get the persistent team that the lineup is playing as, if any.
   */
  team?: GQLTeam;

  /**
   * All members of this lineup.
   */
  members: Array<GQLTournamentLineupMember | null>;

  /**
   * The placement in the tournament for this lineup.
   *
   * If no results are available yet this will be null.
   */
  placement?: GQLTournamentLineupPlacement;

  /**
   * The participation status of the lineup.
   */
  participationStatus: GQLTournamentLineupParticipation;

  /**
   * The ID of the lineup.
   */
  id: GQLUUID;

  /**
   * The name of the tournament lineup. It may be sourced from a Challengermode team or party.
   * The value returned here is a snapshot, but it is not immutable - it might be adjusted when the name of a team changes globally, or it might not.
   */
  name: string;

  /**
   * Denotes if a lineup has been disqualified from the tournament. A disqualified lineup will concede all matches, and disappear from the bracket entirely.
   */
  disqualified: boolean;

  /**
   * A team's tournament seed is its ranking or placement in the tournament bracket, determined by automatic factors or manual input, to ensure fair and balanced matchups.This number is 0-indexed where lower means better. A team will always have if they are part of the rooster and tournament has started.
   */
  seed?: number;

  /**
   * A manual seed value that can be used during manual seeding to set the teams actual seed.
   */
  preSeed?: number;
}
// ...
export interface GQLTournamentMatchContext extends GQLITournamentContext {
  /**
   * The stage number of this match in the tournament.
   */
  stageNumber: number;

  /**
   * The round number of this match in the tournament.
   */
  roundNumber: number;

  /**
   * If relevant, the specific lineup in this context.
   */
  lineup?: GQLMatchLineup;

  /**
   * If relevant, the member specific in this context.
   */
  member?: GQLMatchMember;

  /**
   * Reference to the tournament.
   */
  tournament?: GQLTournament;
}
// ...
export const enum GQLMatchSeriesState {
  UNKNOWN = "UNKNOWN",

  /**
   * The match series has been created and some "parameters" (game settings, teams) are known. Playing has not yet begun, either because parameters are missing or because it needs to be started manually.
   */
  WAITING = "WAITING",

  /**
   * The match series is currently running and some matches have been created.
   */
  RUNNING = "RUNNING",

  /**
   * The match series is paused. If the match involves several games, there may be some time in between each game (to allow for player substitutions, short breaks etc).
   */
  PAUSED = "PAUSED",

  /**
   * The match series is completed and the result of this match has been decided.
   */
  COMPLETED = "COMPLETED",

  /**
   * The result of this match was discarded, and "no opponent" was propagated.
   */
  NULLIFIED = "NULLIFIED",
}
// ...
export const enum GQLTournamentNodeLabel {
  /**
   * The upper bracket of a double elimination stage. Teams that lose in this bracket are sent to the lower bracket.
   */
  DOUBLE_ELIMINATION_UPPER_BRACKET = "DOUBLE_ELIMINATION_UPPER_BRACKET",

  /**
   * The lower bracket of a double elimination stage.Teams that lose in this bracket are eliminated from the tournament.
   */
  DOUBLE_ELIMINATION_LOWER_BRACKET = "DOUBLE_ELIMINATION_LOWER_BRACKET",

  /**
   * The grand finals bracket of a double elimination stage.
   */
  DOUBLE_ELIMINATION_FINALS_BRACKET = "DOUBLE_ELIMINATION_FINALS_BRACKET",

  /**
   * A single elimination bracket.
   */
  SINGLE_ELIMINATION_BRACKET = "SINGLE_ELIMINATION_BRACKET",

  /**
   * A bracket using a swiss format - a non-elimination format where players compete against opponents with similar records each round.
   */
  SWISS_BRACKET = "SWISS_BRACKET",

  /**
   * A bracket with round robin format, where each team or player plays against every other team or player in their group for a specified number of encounters.
   */
  ROUND_ROBIN_BRACKET = "ROUND_ROBIN_BRACKET",

  /**
   * The final match of a tournament bracket.
   */
  FINALS_MATCH = "FINALS_MATCH",

  /**
   * A semi-finals match of a tournament bracket.
   */
  SEMI_FINALS_MATCH = "SEMI_FINALS_MATCH",

  /**
   * The bronze match of a tournament bracket.
   */
  BRONZE_MATCH = "BRONZE_MATCH",

  /**
   * The final round of a tournament bracket.
   */
  FINALS_ROUND = "FINALS_ROUND",

  /**
   * The semi-finals round of a tournament bracket.
   */
  SEMI_FINALS_ROUND = "SEMI_FINALS_ROUND",

  /**
   * The quarter-finals round of a tournament bracket.
   */
  QUARTER_FINALS_ROUND = "QUARTER_FINALS_ROUND",
}
// ...
export type GQLDateTime = any;

export type GQLJSON = any;
// ...
export type GQLUUID = any;